branching of 3.5 and arts 1.5
svn path=/branches/KDE/3.5/kdepim/; revision=438982

+++ karm/test/__lib.sh
@@ -7,7 +7,9 @@ function set_up()
 
   if [ -n "$DCOPID" ]; then dcop $DCOPID KarmDCOPIface quit; fi;
 
-  if [ -e "$TESTFILE" ]; then rm $TESTFILE; fi
+  if [ "x$SKIP_TESTFILE_DELETE" != "xtrue" ]; then
+	  if [ -e "$TESTFILE" ]; then rm $TESTFILE; fi
+  fi
 
   #echo "__lib.sh - starting karm with $TESTFILE"
   karm "$TESTFILE" & 
@@ -47,12 +49,12 @@ function set_up()
     sleep 1
   done
 
-  echo "__lib.sh: DCOPID = $DCOPID, KARM_VERSION = $KARM_VERSION"
-
   if [ "x$DCOPID" = x ]
   then
     echo "__lib.sh set_up error: could not start karm--no dcop id."
     exit 1
+  else
+    echo "__lib.sh: DCOPID = $DCOPID, KARM_VERSION = $KARM_VERSION"
   fi
 
   if [ "x$KARM_VERSION" = x ]
@@ -71,5 +73,7 @@ function tear_down()
 {
   if [ -n "$DCOPID" ]; then dcop "$DCOPID" KarmDCOPIface quit; fi;
 
-  if [ -e "$TESTFILE" ]; then rm "$TESTFILE"; fi
+  if [ "x$SKIP_TESTFILE_DELETE" != "xtrue" ]; then
+		if [ -e "$TESTFILE" ]; then rm "$TESTFILE"; fi
+	fi
 }
